Salvando ProdutoServico

+++ src/main/java/com/dynns/cloudtecnologia/senior/rest/dto/produtoServico/ProdutoServicoNewDTO.java
@@ -10,7 +10,6 @@ import lombok.NoArgsConstructor;
 import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -27,5 +26,5 @@ public class ProdutoServicoNewDTO {
 
     @NotNull(message = "{campo.nomeArquivo.obrigatorio}")
     @DecimalMin(value = "1.00", message = "{campo.preco.maiorQueZero}")
-    private BigDecimal preco;
+    private Double preco;
 }
